boardsはコマンドライン引数でパスを指定されたときのみ出力されるようにした

+++ books/visualize_search_board/mcts_on_book_dl.py
@@ -176,7 +176,7 @@ if __name__ == "__main__":
     args.add_argument('book')
     args.add_argument('dl_pickle')
     args.add_argument('sfens')
-    args.add_argument('--boards', type=str, default='test.pickle')
+    args.add_argument('--boards', type=str)
     args.add_argument('--root_sfens', type=str)
     args.add_argument('--book_moves_threshold', type=int, default=4)
     args.add_argument('--eval_diff', type=int, default=30)
@@ -283,10 +283,12 @@ if __name__ == "__main__":
     moves_list = []
     for _, key in move_count_list:
         sfens_list.append(f"sfen {dl_data_tree[key].board.sfen()}\n")
-        moves_list.append(dl_data_tree[key].board.history)
+        if args.boards and os.path.exists(args.boards):
+            moves_list.append(dl_data_tree[key].board.history)
 
     with open(args.sfens, "w") as f:
         f.writelines(sfens_list)
 
-    with open(args.boards, "wb") as f:
-        pickle.dump(moves_list, f)
+    if args.boards and os.path.exists(args.boards):
+        with open(args.boards, "wb") as f:
+            pickle.dump(moves_list, f)
